feat: Phase 3 backend — Memory Inspector, recs, anonymous consent
/memory (GET/PATCH/DELETE) backs the Inspector: lists every belief regardless of
status plus the full audit trail (a direct DB read, like consolidation/decay —
doesn't fit recall's budget-capped active-only shape), corrects via revise_belief,
deletes via forget audited as "user_delete" (vs system-driven "deprecate").

/recs is the hybrid retrieval endpoint: embeds query ∪ recalled belief statements,
searches the catalog, reranks with qwen-turbo, falls back to similarity-only
ordering (reporting degraded=true) if Qwen is unavailable — this is the exact
endpoint CLAUDE.md's graceful-degradation rule names. Deleting a belief in the
Inspector changes what recall returns, which changes what /recs returns next.

Anonymous shoppers (consent banner declined) now get a real separate path:
/events and /chat take a persist flag, and declining skips Postgres and Qwen
entirely in favor of a TTL'd Redis session log (app/services/session_store.py) —
architecture rule 5 requires opt-in persistence, not opt-out-after-the-fact.

Fixed two real bugs found while building this: MemoryClient collapsed tool-level
rejections (e.g. "belief not found") into the same exception as transport
failures, which would have reported 404s as 503s; and cosine_similarity could
return a numpy float32 (pgvector rows), which broke JSON-encoding the /recs
rerank prompt.

41 tests in apps/api (was 32), 35 in mcp-memory.

+++ packages/mcp-memory/mcp_memory/ranking.py
@@ -34,7 +34,9 @@ def cosine_similarity(a: list[float], b: list[float]) -> float:
     norm_b = math.sqrt(sum(y * y for y in b))
     if norm_a == 0 or norm_b == 0:
         return 0.0
-    return dot / (norm_a * norm_b)
+    # numpy inputs (pgvector rows) make `dot`/`norm_*` numpy scalars (e.g. float32),
+    # which json.dumps can't serialize — always return a plain Python float.
+    return float(dot / (norm_a * norm_b))
 
 
 def recency_weight(days_since_reinforced: float, half_life_days: int = 30) -> float:
